updated code for rfq, extracted duplicate code
