+++ src/components/WalletButton/WalletButton.stories.tsx
@@ -35,7 +35,6 @@ const Template: Story<WalletButtonProps> = (args) => (
 const transactions: SubmittedTransaction[] = [
   {
     type: "Order",
-    nonce: "69",
     order: {
       expiry: "1628840480",
       nonce: "1628840180688",
@@ -56,7 +55,6 @@ const transactions: SubmittedTransaction[] = [
   } as SubmittedTransaction,
   {
     type: "Order",
-    nonce: "69",
     order: {
       expiry: "1628840474",
       nonce: "1628840174080",
@@ -77,7 +75,6 @@ const transactions: SubmittedTransaction[] = [
   } as SubmittedTransaction,
   {
     type: "Order",
-    nonce: "69",
     order: {
       expiry: "1628840464",
       nonce: "1628840164346",
